Removed shield and hull resistance numbers and charts, since these were completely wrong
In order for these be correct, one would have to calculate these values on a per attack basis. It may not be possible to reliably match shield and hull hit of an attack together. So it may not be possible to get these stats correctly in all instances from the log. Therefore this feature will not be added to the analyzer.